Rearrange existing code
